Adds map centering when changing cities
