+++ src/components/host/host.tsx
@@ -33,14 +33,7 @@ function Host({ host, description }: THostProps): React.JSX.Element {
 			</div>
 			<div className="offer__description">
 				<p className="offer__text">
-					{description}.
-					<br />
-					А ниже не совсем понимаю, как быть со вторым параграфом. В примере на сервере вообще одно короткое предложение.
-				</p>
-				<p className="offer__text">
-					An independent House, strategically located between Rembrand
-					Square and National Opera, but where the bustle of the city
-					comes to rest in this alley flowery and colorful.
+					{description}
 				</p>
 			</div>
 		</div>
