+++ src/pages/main-page/main-page.tsx
@@ -1,6 +1,6 @@
 import { Helmet } from 'react-helmet-async';
-import { useEffect, useState } from 'react';
-import { useActionCreators, useAppSelector } from '../../hooks';
+import { useState } from 'react';
+import { useAppSelector } from '../../hooks';
 import classNames from 'classnames';
 import Header from '../../components/header/header';
 import OffersList from '../../components/offers-list/offers-list';
@@ -9,12 +9,8 @@ import CityList from '../../components/city-list/city-list';
 import Spinner from '../../components/spinner/spinner';
 import { getOffersByCity } from '../../utils';
 import { RequestStatus } from '../../const';
-import { useAuth } from '../../hooks/use-authorize';
-import { favoritesActions } from '../../store/slices/favorites';
 
 function MainPage(): React.JSX.Element {
-	const favoriteActions = useActionCreators(favoritesActions);
-	const isAuthorized = useAuth();
 	const offers = useAppSelector((state) => state.OFFERS.offers);
 	const offersByCity = getOffersByCity(offers);
 	const selectedCity = useAppSelector((state) => state.OFFERS.selectedCity);
@@ -22,12 +18,6 @@ function MainPage(): React.JSX.Element {
 	const [id, setId] = useState<null | string>(null);
 	const isLoading = offersFetchingStatus === RequestStatus.Pending;
 	const hasOffers = offersByCity[selectedCity] && offersByCity[selectedCity].length > 0;
-
-	useEffect(() => {
-		if (isAuthorized) {
-			favoriteActions.fetchFavorites();
-		}
-	}, [isAuthorized, favoriteActions]);
 
 	function handleMouseEnter(offerId: string): void {
 		setId(offerId);
